# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jamvi API — personal and group money management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /budget-categories:
    get:
      operationId: getBudgetCategories
      tags: [budget-categories]
      summary: List all budget categories with amounts
      responses:
        "200":
          description: Budget categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetCategory"
    post:
      operationId: createBudgetCategory
      tags: [budget-categories]
      summary: Create a budget category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetCategoryInput"
      responses:
        "201":
          description: Created budget category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategory"
        "400":
          description: Invalid category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A category with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /budget-categories/recommendations/apply:
    post:
      operationId: applyBudgetCategoryRecommendations
      tags: [budget-categories]
      summary: Add missing recommended categories to the active workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetCategoryRecommendationApplyInput"
      responses:
        "200":
          description: Recommendations applied without changing existing categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategoryRecommendations"
  /budget-categories/{id}:
    put:
      operationId: updateBudgetCategory
      tags: [budget-categories]
      summary: Update a budget category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetCategoryUpdateInput"
      responses:
        "200":
          description: Updated budget category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategory"
        "400":
          description: Invalid category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A category with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBudgetCategory
      tags: [budget-categories]
      summary: Delete a budget category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    BudgetCategory:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        budgetAmount:
          type: number
          description: Monthly budget in KES
        priority:
          type: number
          description: 1=survival essentials, 2=health/education, 3=household, 4=connectivity, 5=discretionary
        color:
          type: string
        isRecurring:
          type: boolean
          description: Whether this budget applies every month
        activeMonth:
          type: ["number", "null"]
          description: Month when a one-time budget applies
        activeYear:
          type: ["number", "null"]
          description: Year when a one-time budget applies
      required:
        - id
        - name
        - budgetAmount
        - priority
        - color
        - isRecurring
    BudgetCategoryInput:
      type: object
      properties:
        name:
          type: string
          description: Budget category name. "Uncategorized" is reserved internally and cannot be used.
        budgetAmount:
          type: number
          description: Monthly budget in KES; use 0 when creating a category from a withdrawal
        priority:
          type: number
        color:
          type: string
        isRecurring:
          type: boolean
          default: true
        activeMonth:
          type: ["number", "null"]
          minimum: 1
          maximum: 12
        activeYear:
          type: ["number", "null"]
          minimum: 2000
          maximum: 2200
      required:
        - name
        - budgetAmount
    BudgetCategoryUpdateInput:
      type: object
      description: Fields to update. A one-time budget requires both activeMonth and activeYear.
      properties:
        name:
          type: string
          description: Budget category name. "Uncategorized" is reserved internally and cannot be used.
        budgetAmount:
          type: number
        priority:
          type: number
        color:
          type: string
        isRecurring:
          type: boolean
        activeMonth:
          type: ["number", "null"]
          minimum: 1
          maximum: 12
        activeYear:
          type: ["number", "null"]
          minimum: 2000
          maximum: 2200
    BudgetCategoryRecommendation:
      type: object
      properties:
        name:
          type: string
        budgetAmount:
          type: number
        priority:
          type: number
        color:
          type: string
        exists:
          type: boolean
      required: [name, budgetAmount, priority, color, exists]
    BudgetCategoryRecommendations:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/GroupKind"
        existing:
          type: array
          items:
            $ref: "#/components/schemas/BudgetCategoryRecommendation"
        missing:
          type: array
          items:
            $ref: "#/components/schemas/BudgetCategoryRecommendation"
      required: [kind, existing, missing]
    BudgetCategoryRecommendationApplyInput:
      type: object
      description: Applies only missing recommendations; existing categories are never changed.
      properties:
        confirm:
          type: boolean
          default: true
...
    GroupKind:
      type: string
      enum: [personal, family, chama, club, team, student_group, other]
...